requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autoenv==1.0.0
bcrypt==3.1.7
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.7
entrypoints==0.3
fatsecret==0.2.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.4.5
idna==2.7
importlib-metadata==0.18
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==3.0.1
path.py==12.0.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.2.2
rauth==0.7.3
requests==2.19.1
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.23
virtualenv==16.4.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
zipp==0.5.1
